Consolidate duplicated code.
