+++ radio/src/gui/navigation/navigation_9x.cpp
@@ -19,31 +19,7 @@
  * GNU General Public License for more details.
  */
 
-#include "opentx.h"
-#include "navigation.h"
-
-#include "switches.h"
-
-vertpos_t menuVerticalOffset;
-vertpos_t menuVerticalPosition;
-horzpos_t menuHorizontalPosition;
-
-int8_t s_editMode;
-uint8_t noHighlightCounter;
-uint8_t menuCalibrationState;
-
-int8_t  checkIncDec_Ret;
-
-INIT_STOPS(stops100, 3, -100, 0, 100)
-INIT_STOPS(stops1000, 3, -1000, 0, 1000)
-INIT_STOPS(stopsSwitch, 15, SWSRC_FIRST,
-           CATEGORY_END(-SWSRC_FIRST_LOGICAL_SWITCH),
-           CATEGORY_END(-SWSRC_FIRST_TRIM),
-           CATEGORY_END(-SWSRC_LAST_SWITCH + 1), 0,
-           CATEGORY_END(SWSRC_LAST_SWITCH), CATEGORY_END(SWSRC_FIRST_TRIM - 1),
-           CATEGORY_END(SWSRC_FIRST_LOGICAL_SWITCH - 1), SWSRC_LAST)
-
-extern int checkIncDecSelection;
+#define POS_HORZ_INIT(posVert)         0
 
 int checkIncDec(event_t event, int val, int i_min, int i_max,
                 unsigned int i_flags, IsValueAvailable isValueAvailable,
@@ -148,74 +124,7 @@ int checkIncDec(event_t event, int val, int i_min, int i_max,
     newval = !val;
   }
 
-  if (i_flags & (INCDEC_SOURCE | INCDEC_SOURCE_VALUE)) {
-    if (event == EVT_KEY_LONG(KEY_ENTER)) {
-      checkIncDecSelection = MIXSRC_NONE;
-
-      if (i_flags & INCDEC_SOURCE_VALUE && isSource) {
-        POPUP_MENU_ADD_ITEM(STR_CONSTANT);
-      }
-
-      if (i_min <= MIXSRC_FIRST_INPUT && i_max >= MIXSRC_FIRST_INPUT) {
-        if (getFirstAvailable(MIXSRC_FIRST_INPUT, MIXSRC_LAST_INPUT, isInputAvailable) != MIXSRC_NONE &&
-            (i_flags & INCDEC_SOURCE_NOINPUTS) == 0) {
-          POPUP_MENU_ADD_ITEM(STR_MENU_INPUTS);
-        }
-      }
-#if defined(LUA_MODEL_SCRIPTS)
-      if (i_min <= MIXSRC_FIRST_LUA && i_max >= MIXSRC_FIRST_LUA) {
-        if (getFirstAvailable(MIXSRC_FIRST_LUA, MIXSRC_LAST_LUA, isSourceAvailable) != MIXSRC_NONE) {
-          POPUP_MENU_ADD_ITEM(STR_MENU_LUA);
-        }
-      }
-#endif
-      if (i_min <= MIXSRC_FIRST_STICK && i_max >= MIXSRC_FIRST_STICK)      POPUP_MENU_ADD_ITEM(STR_MENU_STICKS);
-      if (i_min <= MIXSRC_FIRST_POT && i_max >= MIXSRC_FIRST_POT)          POPUP_MENU_ADD_ITEM(STR_MENU_POTS);
-      if (i_min <= MIXSRC_MIN && i_max >= MIXSRC_MIN)                      POPUP_MENU_ADD_ITEM(STR_MENU_MIN);
-      if (i_min <= MIXSRC_MAX && i_max >= MIXSRC_MAX)                      POPUP_MENU_ADD_ITEM(STR_MENU_MAX);
-#if defined(HELI)
-      if (modelHeliEnabled())
-        if (i_min <= MIXSRC_FIRST_HELI && i_max >= MIXSRC_FIRST_HELI && isValueAvailable && isValueAvailable(MIXSRC_FIRST_HELI))
-          POPUP_MENU_ADD_ITEM(STR_MENU_HELI);
-#endif
-      if (i_min <= MIXSRC_FIRST_TRIM && i_max >= MIXSRC_FIRST_TRIM)        POPUP_MENU_ADD_ITEM(STR_MENU_TRIMS);
-      if (i_min <= MIXSRC_FIRST_SWITCH && i_max >= MIXSRC_FIRST_SWITCH)    POPUP_MENU_ADD_ITEM(STR_MENU_SWITCHES);
-      if (i_min <= MIXSRC_FIRST_TRAINER && i_max >= MIXSRC_FIRST_TRAINER)  POPUP_MENU_ADD_ITEM(STR_MENU_TRAINER);
-      if (i_min <= MIXSRC_FIRST_CH && i_max >= MIXSRC_FIRST_CH)            POPUP_MENU_ADD_ITEM(STR_MENU_CHANNELS);
-      if (i_min <= MIXSRC_FIRST_GVAR && i_max >= MIXSRC_FIRST_GVAR && isValueAvailable && isValueAvailable(MIXSRC_FIRST_GVAR)) {
-        POPUP_MENU_ADD_ITEM(STR_MENU_GVARS);
-      }
-
-      if (modelTelemetryEnabled() && i_min <= MIXSRC_FIRST_TELEM && i_max >= MIXSRC_FIRST_TELEM) {
-        for (int i = 0; i < MAX_TELEMETRY_SENSORS; i++) {
-          TelemetrySensor * sensor = & g_model.telemetrySensors[i];
-          if (sensor->isAvailable()) {
-            POPUP_MENU_ADD_ITEM(STR_MENU_TELEMETRY);
-            break;
-          }
-        }
-      }
-      if (i_flags & INCDEC_SOURCE_INVERT) POPUP_MENU_ADD_ITEM(STR_MENU_INVERT);
-      POPUP_MENU_START(onSourceLongEnterPress);
-    }
-    if (checkIncDecSelection != 0) {
-      if (i_flags & INCDEC_SOURCE_VALUE) {
-        if (checkIncDecSelection == MIXSRC_VALUE) {
-          newval = 0;
-          isSource = false;
-        } else {
-          newval = (checkIncDecSelection == MIXSRC_INVERT ? -newval : checkIncDecSelection);
-          if (checkIncDecSelection != MIXSRC_INVERT)
-            isSource = true;
-        }
-      } else {
-        newval = (checkIncDecSelection == MIXSRC_INVERT ? -newval : checkIncDecSelection);
-      }
-      if (checkIncDecSelection != MIXSRC_MIN && checkIncDecSelection != MIXSRC_MAX)
-        s_editMode = EDIT_MODIFY_FIELD;
-      checkIncDecSelection = 0;
-    }
-  }
+  newval = showPopupMenus(event, newval, i_min, i_max, i_flags, isValueAvailable);
 
   if (newval != val) {
     if (!(i_flags & NO_INCDEC_MARKS) && (newval != i_max) &&
@@ -244,50 +153,6 @@ int checkIncDec(event_t event, int val, int i_min, int i_max,
   }
 
   return newval;
-}
-
-#define SCROLL_POT1_TH 32
-
-#define CURSOR_NOT_ALLOWED_IN_ROW(row) ((int8_t)MAXCOL(row) < 0)
-
-#define INC(val, min, max)             if (val<max) {val++;} else {val=min;}
-#define DEC(val, min, max)             if (val>min) {val--;} else {val=max;}
-
-tmr10ms_t menuEntryTime;
-
-#define MAXCOL_RAW(row)                (horTab ? *(horTab+min(row, (vertpos_t)horTabMax)) : (const uint8_t)0)
-#define MAXCOL(row)                    (MAXCOL_RAW(row) >= HIDDEN_ROW ? MAXCOL_RAW(row) : (const uint8_t)(MAXCOL_RAW(row) & (~NAVIGATION_LINE_BY_LINE)))
-#define POS_HORZ_INIT(posVert)         0
-
-uint8_t chgMenu(uint8_t curr, const MenuHandler * menuTab, uint8_t menuTabSize, int direction)
-{
-  int cc = curr + direction;
-  while (cc != curr) {
-    if (cc < 0)
-      cc = menuTabSize - 1;
-    else if (cc >= menuTabSize)
-      cc = 0;
-    if (menuTab[cc].isEnabled())
-      return cc;
-    cc += direction;
-  }
-  return curr;
-}
-
-uint8_t menuSize(const MenuHandler * menuTab, uint8_t menuTabSize)
-{
-  uint8_t sz = 0;
-  for (int i = 0; i < menuTabSize; i += 1) {
-    if (menuTab[i].isEnabled()) {
-      sz += 1;
-    }
-  }
-  return sz;
-}
-
-uint8_t menuIdx(const MenuHandler * menuTab, uint8_t curr)
-{
-  return menuSize(menuTab, curr + 1) - 1;
 }
 
 void check(event_t event, uint8_t curr, const MenuHandler *menuTab,
